cmd/tx-storm: txsDict is moved to meta.Txs for imports
